Edit Equipment 와 겹치는 부분 ModifyEquipment로 분리
